Tests, stability and config
Improved overall stability and switched my unit
testing framework. This also adds some
functionality around configuration of various
keybinding settings that can't/shohudln't be
stored in keybindings.json. Unforuntately this bit
has yet to be tested, but there's so much changed
that I'd like to move on to a new branch.