Bugfix: profile switching
The prior implementation didn't update relevant
settings when the user switches profiles. This
implementation synchronizes the config state with
a name + hash of the bindings file that is stored
in user's preferences. This naturally works with
anything that changes the config and the
keybindings together.

It does mean the user can change the config
manually and it will be out of sync with the
keybindings, so I have placed a warning that the
user should use appropriate commands instead.