Fix bug in markdown documentation generation (#8)
Release 0.2.0 had a bug in the final implementation that was not caught by unit tests, but did show up for Larkin. This bug removed all binding tables from Larkin's markdown documentation. The root cause was that the older implementation assumed all bindings would be provided in the order of their index, which was not the case. They are sorted by `priority`.

This PR implements moves to an easier to follow implementation that 1.) indexes all resolved bindings by their `index` and 2.) includes an index field for documentation bindings. We can get then straightforwardly match resolved bindings to documentation bindings.