Implement `fallbackBindings` for `mode` specification (#40)
This implements a `fallbackBinding` field. It can be set to the name of
an existing mode. It causes any unspecified keys for the given mode
fallback to the bindings provided in the fallback. A limitation of the
current implementation is that to work properly, the bindings for this
mode must be defined before the fallback bindings.