Eliminate `lineNumbers` mode option (#61)
Changing how `lineNumbers` are displayed per mode seemed to work well in
older versions of VSCode, but it is now causing unexpected cursor
movement whenever the mode is switched. This is a pretty unacceptable
problem for day-to-day use, so I'm removing this feature for now. If at
some point VSCode makes it possible to change line number display
without changing the config file, I can reconsider this feature.